+++ src/pages/quarantine.tsx
@@ -1,25 +1,23 @@
 import { AppLayout } from "@/components/layout";
 import { ShieldCheck } from "lucide-react"
 import { Empty, EmptyDescription, EmptyHeader, EmptyMedia, EmptyTitle } from "@/components/ui/empty";
-import { lazy, Suspense, useEffect, useState, useTransition } from "react";
-import { IQuarantineData } from "@/lib/types";
+import { lazy, Suspense, useEffect, useMemo, useState, useTransition } from "react";
 import { invoke } from "@tauri-apps/api/core";
-import { formatBytes } from "@/lib/helpers";
 import Popup from "@/components/popup";
 import { toast } from "sonner";
 import { IQuarantineState } from "@/lib/types/states";
 import { INITIAL_QUARANTINE_STATE } from "@/lib/constants/states";
 import QuarantineLoader from "@/loaders/quarantine";
+import { IQuarantineData } from "@/lib/types";
+import { formatBytes } from "@/lib/helpers";
+import useSettings from "@/hooks/use-settings";
+import { GET_QUARANTINE_COLS } from "@/components/data-table/columns/quarantine";
 const QuarantineTable = lazy(()=>import("@/contents/quarantine"))
 
 export default function QuarantinePage(){
-     const [isRefreshing, startTransition] = useTransition();
+     const {settings} = useSettings();
      const [quarantineState, setQuarantineState] = useState<IQuarantineState>(INITIAL_QUARANTINE_STATE);
-     const setState = (overrides: Partial<IQuarantineState>) =>
-          setQuarantineState(prev=>({
-               ...prev,
-               ...overrides
-          }))
+     const [isRefreshing, startTransition] = useTransition();
      const fetchData = () => {
           startTransition(()=>invoke<IQuarantineData<"type">[]>("list_quarantine").then(data=>{
                const newData: IQuarantineData<"state">[] = data.map(({id,threat_name,file_path,quarantined_at,size})=>({
@@ -32,9 +30,8 @@ export default function QuarantinePage(){
                setState({ data: newData });
           }).catch(() => setState({ data: [] })));
      }
-     useEffect(()=>{
-          fetchData()
-     },[])
+     useEffect(()=>fetchData(),[])
+     const setState = (overrides: Partial<IQuarantineState>) => setQuarantineState(prev=>({ ...prev, ...overrides }))
      const quarantineAction = async(type: "restore" | "delete") => {
           const commandName = `${type}_quarantine`
           const message = type==="restore" ? "File restored from quarantine!" : "The file has been deleted permanently";
@@ -58,41 +55,36 @@ export default function QuarantinePage(){
                })
           }
      }
-     const handleBulkDelete = async()=>{
-          setState({bulkDelete: false})
-          try {
+     const bulkAction = async(type: "restore" | "delete") => {
+          const key = type==="restore" ? "bulkRestore" : "bulkDelete"
+          setState({ [key]: false })
+          try{
                const ids = data.map(t => t.id);
-               await invoke("clear_quarantine", { ids });
+               const commandName = type==="restore" ? "restore_all" : "clear_quarantine";
+               const msgName = type === "restore" ? "All Threats restored" : "All Threats deleted"
+               await invoke(commandName, { ids });
                setState({ data: [] })
-               toast.success("All threats deleted");
-          } catch {
-               toast.error("Failed to delete all threats");
-          }
-     }
-     const handleBulkRestore = async()=>{
-          setState({
-               bulkRestore: false
-          })
-          try {
-               const ids = data.map(t => t.id);
-               await invoke("restore_all", { ids });
-               setState({ data: [] })
-               toast.success("All threats restored!");
-          } catch {
-               toast.error("Failed to restore all threats");
+               toast.success(msgName);
+          } catch (err){
+               const errMsg = type==="restore" ? "Failed to restore all threats" : "Failed to delete all threats";
+               toast.error(errMsg);
+               console.error(err)
           }
      }
      const {isOpenDelete, isOpenRestore, bulkDelete, bulkRestore, data} = quarantineState
+     const isNotEmpty = useMemo(()=>data.length>0,[data])
      return (
           <AppLayout className="flex justify-center items-center gap-4 flex-col p-4">
                <h1 className="text-2xl md:text-3xl lg:text-4xl font-medium border-b pb-2 w-fit">Quarantine</h1>
-               {data.length>0 ? (
-                    <Suspense fallback={<QuarantineLoader/>}>
+               {isNotEmpty ? (
+                    <Suspense fallback={<QuarantineLoader rows={data.slice(0,10).length}/>}>
                          <QuarantineTable
-                              isRefreshing={isRefreshing}
-                              fetchData={fetchData}
-                              setState={setState}
                               data={data}
+                              isRefreshing={isRefreshing}
+                              onRefresh={fetchData}
+                              onBulkClear={()=>setState({ bulkDelete: true })}
+                              onBulkRestore={()=>setState({ bulkRestore: true })}
+                              columns={GET_QUARANTINE_COLS(setState,settings.developerMode)}
                          />
                     </Suspense>
                ) : (
@@ -130,7 +122,7 @@ export default function QuarantinePage(){
                     description="Continue?"
                     submitTxt="Delete"
                     closeText="Cancel"
-                    submitEvent={handleBulkDelete}
+                    submitEvent={()=>bulkAction("delete")}
                />
                <Popup
                     open={bulkRestore}
@@ -139,7 +131,7 @@ export default function QuarantinePage(){
                     description="Continue?"
                     submitTxt="Restore"
                     closeText="Cancel"
-                    submitEvent={handleBulkRestore}
+                    submitEvent={()=>bulkAction("restore")}
                />
           </AppLayout>
      )